Progress report 1 on Frontpage-gui
A lot of the HTML and CSS is in place, there is no functionality yet. There is established an connection to Font Awesome's library for icon usage.
There is currently lorum ipsum text in place of what should be there, just to keep tabs on how the html is placed.